+++ HISTORY.md
@@ -17,6 +17,7 @@
 
 ### Added:
 
+- Basic JSON output for workflows and tasks aborted by OPH_CANCEL
 - Argument 'exec_mode' to OPH_WAIT
 - Monitor of the number of active threads in server [#53](https://github.com/OphidiaBigData/ophidia-server/pull/53)
 - Options to generate random user passwords and execute post-scripts to set env variables [#52](https://github.com/OphidiaBigData/ophidia-server/pull/52)